added a image slider at the bottom of website.
This took forever for some reason for me to figure out but finally here
it is and I can easily change the dimensions of the images, scroller
size, scoller speed. and margins depending what you want. Also can
easily add a hover function for when the mouse is over the images the
automatically scroll will stop