Changed the slideshow into a image fader
I spent a lot of time trying to tweek the previous image slideshow i
created to make so it would scroll for infinity and i could not figure
out how to make it work. I commented out the previous slider work just
in case you or I could figure it out. I decided though that this image
fader was pretty cool and it infinitly loops through whatever pictures
you put in the loop so I went with it.